requirements.txt
fastapi[standard]
mail-parser
uvicorn
python-telegram-bot
python-dotenv
sqlalchemy[asyncio]
aiosqlite


